rename: 패키지 변경이 허용된 클래스 폴더 구조 변경
